utilityとdescriptor_heapの分離
