+++ ProjectDirectX12/ProjectDirectX12.vcxproj
@@ -144,7 +144,9 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="create_view.cpp" />
+    <ClCompile Include="descriptor_heap.cpp" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="utility.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="command_manager.hpp" />
