+++ ProjectDirectX12/ProjectDirectX12.vcxproj.filters
@@ -3,6 +3,8 @@
   <ItemGroup>
     <ClCompile Include="main.cpp" />
     <ClCompile Include="create_view.cpp" />
+    <ClCompile Include="descriptor_heap.cpp" />
+    <ClCompile Include="utility.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="window.hpp" />
